+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-### Fixed
- - Palette Editor: Updated color correction to match Sameboy 0.15.4 "Modern - Balanced" and "Harsh Reality" settings. Renamed color space settings to hopefully make them easier to understand.
+### Changed
+ - Palette Editor: Updated color correction to match Sameboy 0.15.4.
 
 ## [1.13.0] - 2022-04-11
 ### Fixed